backend/requirements.txt
# Municipal Chatbot Backend Dependencies
# Self-hosted stack - no external API dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# LLM & RAG (Ollama - local inference)
langchain>=0.2.0
langchain-ollama>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0

# Vector Database
chromadb>=0.4.0

# Content Ingestion (for RAG pipeline)
beautifulsoup4>=4.12.0
lxml>=5.1.0
requests>=2.31.0
pymupdf>=1.23.0

# Utilities
python-json-logger>=2.0.0